requirements.txt
flask==2.2.5
python-dotenv==0.21.0
requests